Changes on calendarEvents.js
Changed code related to revealDisplayC(), now it doesn't return a string for display value, it returns a boolean value. The use of that value has been modified as well (line 167) from a single line that set the display value to an if conditional that removes a CSS class named "hide" that sets the default display value of the container that holds the id to "none".
Changed icon on getData() for month case 9, day case 29. It showed the icon for "Inktober" instead of "Change to winter time".