New December logic
In Logic and Moons, added a conditional that uses a "storedCurrentYear", so the calendar can be updated in December and use the current year's events while next year it will use the new ones, so it doesn't have to be updated twice (first in December, keeping that year's events and then in January changing December events to the ones of the new year. Now both can be stored and the calendar will showcase a list or another depending on the year, "automating" the update and needing just one).